+++ Assets/Global/Scripts/Menu/Loader.cs
@@ -107,6 +107,9 @@ public class Loader : MonoBehaviour
             
             // reset prevLevel after initiating load of previous level
             PreviousLevel.Instance.ResetLevelForRetry();
+
+            GlobalReference.AttemptInvoke(Events.SPEEDRUN_MODE_ACTIVE);
+            
         }
         else
         {
@@ -129,7 +132,6 @@ public class Loader : MonoBehaviour
     private Phase CompleteLoading()
     {
         Time.timeScale = 1;
-        GlobalReference.GetReference<GameManagerReference>().ResetTimers();
         SceneManager.UnloadSceneAsync("Loading");
         GlobalReference.AttemptInvoke(Events.INPUT_ACKNOWLEDGE);
         GlobalReference.AttemptInvoke(Events.SPEEDRUN_MODE_ACTIVE);
