+++ Assets/Global/Scripts/Menu/SpeedrunResultLogic.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
@@ -25,12 +24,22 @@ public class SpeedrunResultLogic : MonoBehaviour
         level2Text.text = GlobalReference.Statistics.Get<string>("level_2_time");
         level3Text.text = GlobalReference.Statistics.Get<string>("level_3_time");
         totalTimeText.text = GlobalReference.Statistics.Get<string>("total_time");
-        totalDeathsText.text = GlobalReference.Statistics.Get<string>("deaths");
+        totalDeathsText.text = GlobalReference.Statistics.Get<int>("deaths").ToString();
     }
 
     public void OnExit() 
     {
         UILogic.FadeToScene("Menu", fadeImage, this);
-        GlobalReference.GetReference<GameManagerReference>().ResetTimers();
+        ResetTimers();
+    }
+
+    void ResetTimers() {
+        GlobalReference.Statistics.Set("level_1_time", "00:00:00");
+        GlobalReference.Statistics.Set("level_2_time", "00:00:00");
+        GlobalReference.Statistics.Set("level_3_time", "00:00:00");
+        GlobalReference.Statistics.Set("total_time", "00:00:00");
+        GlobalReference.Statistics.Set("deaths", 0);
+
+        GlobalReference.Statistics.SaveAll();
     }
 }
